Added Services to CarServiceCreate, added validations if user is disabled

+++ SP/src/Controller/ServiceController.php
@@ -61,25 +61,27 @@ class ServiceController extends Controller
      */
    public function EditService($id, Request $request)
    {
-       $service = $this->getDoctrine()->getRepository(Service::class)->findOneBy(['id' => $id]);
-       $em = $this->getDoctrine()->getManager();
+       if($this->getUser()->getRole() == 2) {
+           $service = $this->getDoctrine()->getRepository(Service::class)->findOneBy(['id' => $id]);
+           $em = $this->getDoctrine()->getManager();
 
-       if(!$service){
-           return $this->render(
-               'profile/index.html.twig',
-               array('error' => 'No service with this ID found',
-               ));
-       }
+           if (!$service) {
+               return $this->render(
+                   'profile/index.html.twig',
+                   array('error' => 'No service with this ID found',
+                   ));
+           }
 
-       $form = $this->createForm(ServiceType::class, $service);
-       $form->handleRequest($request);
+           $form = $this->createForm(ServiceType::class, $service);
+           $form->handleRequest($request);
 
-       if($form->isSubmitted() && $form->isValid()){
-           $getTotalPrice = $service->getPrice() * ($service->getDiscount() / 100);
-           $service->setTotalPrice($service->getPrice() - $getTotalPrice);
-           $em->flush();
+           if ($form->isSubmitted() && $form->isValid()) {
+               $getTotalPrice = $service->getPrice() * ($service->getDiscount() / 100);
+               $service->setTotalPrice($service->getPrice() - $getTotalPrice);
+               $em->flush();
 
-           return $this->redirectToRoute('service');
+               return $this->redirectToRoute('service');
+           }
        }
 
        return $this->render(
@@ -95,6 +97,12 @@ class ServiceController extends Controller
      */
    public function ShowServices()
    {
+       if($this->getUser()->getisDisabled() == 1){
+           return $this->render(
+               'main/index.html.twig',
+               array('error' => 'Your account is disabled. Please contact administrator for more information',
+               ));
+       }
        $service = $this->getDoctrine()->getRepository(Service::class)->findAll();
 
        return $this->render('service/index.html.twig', [
@@ -107,11 +115,13 @@ class ServiceController extends Controller
      */
    public function DeleteService($id, Request $request)
    {
-       $service = $this->getDoctrine()->getRepository(Service::class)->findOneBy(['id' => $id]);
-       if($service) {
-           $em = $this->getDoctrine()->getManager();
-           $em->remove($service);
-           $em->flush();
+       if($this->getUser()->getRoles() == 2) {
+           $service = $this->getDoctrine()->getRepository(Service::class)->findOneBy(['id' => $id]);
+           if ($service) {
+               $em = $this->getDoctrine()->getManager();
+               $em->remove($service);
+               $em->flush();
+           }
        }
        return $this->redirectToRoute('service');
    }
